requirements/prod.txt
# Everything needed in production

# Flask
click>=7.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
requests==2.32.3
pytz==2024.1
# Werkzeug==3.0.3
Flask-Mail==0.10.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
alembic-postgresql-enum==1.3.0

# Migrations
Flask-Migrate==4.0.7

# Forms
# email-validator==2.2.0
# Flask-WTF==1.2.1
# WTForms==3.1.2
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0

# Deployment
# gevent==24.2.1
# gunicorn>=19.9.0
# supervisor==4.2.5

# Flask Static Digest
# Flask-Static-Digest==0.4.1

# Caching & backgroud task
Flask-Caching>=2.0.2
redis==5.0.8
celery==5.4.0

# Debug toolbar
Flask-DebugToolbar==0.15.1

# Environment variable parsing
environs==11.0.0

# setry 
sentry-sdk[flask]==2.13.0
